refactor: files and folder structure updated
