+++ src/constant/general.ts
