--- src/dto/index.ts
@@ -1,4 +1,0 @@
-export interface createCourse {
-  name: string;
-  description: string;
-}
