+++ src/index.tsx
@@ -1,10 +1,10 @@
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
-import App from './App';
 import { Provider } from 'react-redux';
 import { store } from './store';
 import Auth0ProviderWithHistory from './auth/auth0-provider-with-history';
 import './index.css';
+import Navigation from './navigation';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
@@ -13,7 +13,7 @@ root.render(
   <BrowserRouter>
     <Provider store={store()}>
       <Auth0ProviderWithHistory>
-        <App />
+        <Navigation />
       </Auth0ProviderWithHistory>
     </Provider>
   </BrowserRouter>
