+++ src/store/index.ts
@@ -1,4 +1,7 @@
 import { configureStore, combineReducers } from '@reduxjs/toolkit';
+import { useDispatch, useSelector } from "react-redux";
+import type { TypedUseSelectorHook } from "react-redux";
+
 import { API } from './reducer/api';
 import type { PreloadedState } from '@reduxjs/toolkit';
 import { errorLogger } from './middleware/errorLogger';
@@ -25,3 +28,6 @@ export const store = (preloadedState?: PreloadedState<RootState>) => {
 export type RootState = ReturnType<typeof rootReducer>;
 export type AppStore = ReturnType<typeof store>;
 export type AppDispatch = AppStore['dispatch'];
+
+export const useAppDispatch: () => AppDispatch = useDispatch;
+export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
