+++ src/store/reducer/error.ts
@@ -1,6 +1,5 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import _ from 'lodash';
-import { CourseType, ErrorSliceType } from '../../utils/types';
+import { ErrorSliceType } from '../../constant/interface';
 
 const initialState: ErrorSliceType = {
   hasError: false,
